refactor: lecturePlayer->lecture 로 변수명 변경
